+++ src/app/core/services/common/accounting-export.service.ts
@@ -35,19 +35,16 @@ export class AccountingExportService {
     return this.apiService.get(`/workspaces/${this.workspaceId}/accounting_exports/count`, apiParams);
   }
 
-  getAccountingExports(status: AccountingExportStatus[], exportableAccountingExportIds: number[] | null, limit?: number, offset?: number, selectedDateFilter? : SelectedDateFilter | null): Observable<AccountingExportResponse> {
+  getAccountingExports(status: AccountingExportStatus[], exportableAccountingExportIds: number[] | null, limit: number, offset: number, selectedDateFilter? : SelectedDateFilter | null): Observable<AccountingExportResponse> {
     const apiParams: AccountingExportGetParam = {
       type__in: [AccountingExportType.DIRECT_COSTS, AccountingExportType.PURCHASE_INVOICE],
-      status__in: status
+      status__in: status,
+      limit: limit,
+      offset: offset
     };
 
     if (exportableAccountingExportIds?.length) {
       apiParams.id__in = exportableAccountingExportIds;
-    }
-
-    if (limit&&offset) {
-      apiParams.limit = limit,
-      apiParams.offset = offset;
     }
 
     if (selectedDateFilter) {
