+++ src/app/integrations/sage300/sage300-main/sage300-export-log/sage300-complete-export-log/sage300-complete-export-log.component.html
@@ -28,7 +28,7 @@
 
         <app-zero-state-with-illustration *ngIf="!filteredAccountingExports.length"
             [mainText]="'Sorry, no results found!'"
-            [subText]="'We could not find what you were looking for.Kindly check the keywords again.'">
+            [subText]="'We could not find what you were looking for. Kindly check the keywords again.'">
         </app-zero-state-with-illustration>
 
         <app-zero-state-with-illustration *ngIf="totalCount === 0"
