+++ src/app/integrations/sage300/sage300-main/sage300-export-log/sage300-complete-export-log/sage300-complete-export-log.component.ts
@@ -7,7 +7,7 @@ import { Paginator } from 'src/app/core/models/misc/paginator.model';
 import { TrackingService } from 'src/app/core/services/integration/tracking.service';
 import { PaginatorService } from 'src/app/core/services/si/si-core/paginator.service';
 import { environment } from 'src/environments/environment';
-import { AccountingExportList, AccountingExportUtils } from 'src/app/core/models/db/accounting-export.model';
+import { AccountingExportList, AccountingExportClass } from 'src/app/core/models/db/accounting-export.model';
 import { Sage300AccountingExport } from 'src/app/core/models/sage300/db/sage300-accounting-export.model';
 import { ExportLogService } from 'src/app/core/services/common/export-log.service';
 import { WindowService } from 'src/app/core/services/common/window.service';
@@ -34,7 +34,7 @@ export class Sage300CompleteExportLogComponent implements OnInit {
 
   currentPage: number = 1;
 
-  dateOptions: DateFilter[] = ExportLogService.getDateOptions();
+  dateOptions: DateFilter[] = AccountingExportClass.getDateOptions();
 
   selectedDateFilter: SelectedDateFilter | null;
 
@@ -60,14 +60,14 @@ export class Sage300CompleteExportLogComponent implements OnInit {
   ) { }
 
   openExpenseinFyle(expense_id: string) {
-    this.windowService.openInNewTab(ExportLogService.getFyleExpenseUrl(expense_id));
+    this.windowService.openInNewTab(AccountingExportClass.getFyleExpenseUrl(expense_id));
   }
 
   public handleSimpleSearch(event: any) {
     const query = event.target.value.toLowerCase();
 
     this.filteredAccountingExports = this.accountingExports.filter((group: AccountingExportList) => {
-      return ExportLogService.getfilteredAccountingExports(query, group);
+      return AccountingExportClass.getfilteredAccountingExports(query, group);
     });
   }
 
@@ -104,7 +104,7 @@ export class Sage300CompleteExportLogComponent implements OnInit {
           this.totalCount = accountingExportResponse.count;
         }
         const accountingExports: AccountingExportList[] = accountingExportResponse.results.map(accountingExport =>
-          AccountingExportUtils.createAccountingExport(accountingExport, this.exportLogService)
+          AccountingExportClass.parseAPIResponseToExportLog(accountingExport, this.exportLogService)
         );
         this.filteredAccountingExports = accountingExports;
         this.accountingExports = [...this.filteredAccountingExports];
